Starting to implement board hashing.
Wrote a Python script genhash.py to generate a table
of random values for chess position hashing.
The table allows for all the things that are needed:
128 random bits for each piece at each square on the board,
whether or not it is White's turn to move, en passant target files,
and White/Black castling abilities.

However, only the piece positions part of the table is currently used.
I will finish the other stuff in another commit.

PushMove() and PopMove() both update the hash values incrementally,
with a sanity check in PopMove() against a copy of the hash values
stored in the unmove info. This should help catch any bugs with
corrupting the board state.