Was not setting check/immobile flags on underpromotion moves.
I added an extra sanity check in Board.MoveNotation() to make
sure the check/immobile flags were valid inside each formatted move.
This revealed that promotions of a pawn to rook, bishop, or knight
did not have the flags set because of a bug in my optimization
for skipping redundant legality checks.

So I reworked the optimization to still call AddMove(), where those
flags are calculated, but with an extra flag passed in letting
AddMove() know the move has already been proven to be legal.