Finished implementing the hash function of the entire game state.
The hash function now includes all the factors that affect the game state:
- Whether it is White's turn or Black's turn.
- The positions of all the pieces on the board.
- The 4 castling availability flags (WK, WQ, BK, BQ).
- The file a..h of any pawn that can be *legally* captured en passant.

The en passant case required adding checks to determine whether an
en passant capture actually exists and is legal (doesn't put the
capturing side in check), not just that a pawn just moved two
squares forward.

I added an additional array of 16x2 hash salt values for the 2^4
possible castling states, for the sake of calculation speed.