More work on generating and parsing PGN.
Added two new factory functions:
Board.FromPgnText  : creates a board from the first game in a PGN string.
Board.FromGame     : creates a board from a game state.

Instead of saving/preserving state while mutating GameTags,
added a GameTags.Clone() method that makes a copy of an existing object.

When parsing a sequence of chess games from PGN input,
tolerate the final game lacking a result marker;
go ahead and enumerate the final game instead of dropping it on the floor.
However, if we were in the middle of parsing a tag, report a syntax error.

Added better syntax checking of tag names.
Eliminated use of regular expressions for better performance.