Improved move ordering: small bonus for check, BIG bonus for checkmate.
Since I already know whether a move causes check or checkmate,
it makes a lot of sense to use it in move ordering.
Give HUGE bonus for checkmate to bump it toward the front
of the list, and about 0.1 pawn advantage for any move that
causes check. Checking moves tend to constrain what the opponent
can do.