Improved move ordering.
In NegaMax, if there is a move that causes immediate checkmate,
there is no point in searching other moves. Just immediately
return a checkmate score adjusted for depth+1.

Also, always score stalemate moves as having a value of 0 (draw).

Give a slight preference for capturing with smaller pieces,
assuming that they are at risk for recapture.