Detect draws by repetition.
Board.GetGameResult() now finds draw by repetition.

Thinker calls Board.GetGameResult() to find the end of game,
instead of having separate logic. That way, the search understands
draw by repetition.
May need to revisit this part so that we gradually ramp down
the scores of nodes as we approach repetition, instead of
having a horizon cliff in the search.