Adjust scores one ply at a time as they work their way up the tree.
I realized I had an incoherent way of adjusting scores for ply delays.
I was adjusting the scores at the leaf nodes, but that is incorrect
because they end up in the hash table that way. They need to be
adjusted gradually, one ply at a time, as they work their way
up the search tree.

Unfortunately, this causes significantly less efficient pruning,
but that is inherent in anything that spreads out the score values.