Fixed #9 - Treat repeats of positions higher in the tree as draws.
The idea is that we don't want to go in circles.
If any position below the second level is a repeat,
pretend like it is an immediate draw, so that we don't
keep recursing through the loop.
This motivates the thinker to avoid repeats in a
winning position, and try to force repeats in a losing position.
Will still detect genuine draws by repetition at the top level of the tree.