Fixed inconsistent board hashes. Added piece inventory.
I discovered that I was not resetting the pieceHash
when I started over with a new board position.
This means the same board position could have two different
hash values when a Board object was reused.
Now reset pieceHash to (0, 0) every time Board.SetPosition()
is called. I need to remember to initialize everything
related to the board state in SetPosition().

Added an inventory array that is kept up to date with
a tally of each kind of piece on the board.
This speeds up my super-simple material evaluator.
More significantly, it will make things like checking
for insufficient mating material simpler to implement.