Studying which combinations of endgame material can possibly cause checkmate.
As part of figuring out a rock-solid algorithm for Board.GetGameResult()
to handle deciding whether a position is a game-ender or not,
I'm focusing on finding draws due to insufficient material.

There are two levels of draws:

1. Absolute Draws
   I need an extremely conservative definition of a draw
   for the sake of terminating a game between two players.
   If checkmate is theoretically possible through any
   conceivable sequence of legal moves, GetGameResult()
   must return InProgress. Otherwise, acting as an arbiter
   between two human players, this code would prematurely
   terminate the game.

2. Effective Draws
   For the sake of the search engine (the Thinker class)
   I will want a looser definition of a draw, meaning that
   ideal play does not lead to a forced checkmate.

The distinction already occurs in how NegaMax calls RepCount()
to find effective draws, after GetGameResult() has returned InProgress.
I will likely want an additional function for determining
effective draws. Eventually, that function would need access
to a full-blown endgame database to be completely accurate.

So I have created a project PossibleMates. It finds all legal
configurations of two kings and a small set of other pieces,
to find whether there is a possible checkmate. I was already
surprised to find that (unforced) checkmate can occur in K+B vs K+B.

PossibleMates serves as an initial exploration of an endgame
database generator. It motivated me to create some "unsafe"
member functions in the Board class. Use of these new functions
can leave the Board object in a corrupted state.