Refactoring for cleaner Board initialization.
I realized I should make Board.SetPosition() call Board.Clear()
so that the inner state is completely wiped clean each time.
That allows me to have stricter checking in Board.Drop();
we should never drop a piece onto anything but an empty square.
When a capture takes place, the caller first must call Lift()
to remove the captured piece.

Because SetPosition calls Clear, SetPosition can omit resetting
many of the member variables, because Clear does that.

Reworked Board.Clear() to initialize members in the same order
they are declared. This makes the code easier to audit for correctness.

Removed unnecessary code from the unsafe versions of Lift and Drop.

Added PossibleMates to my testall script, so that it is built and run
every time by Travis CI when I push to GitHub.