Board class now has faster (and even more dangerous) utility methods.
Provided more efficient ways to tweak the innards of the Board
object, for the sake of carefully crafted endgame solvers.
Allow direct access to the squares array, and provide utility
methods PlaceWhiteKing and PlaceBlackKing that keep king offsets updated.

Board.bkofs and Board.wkofs are now set to -1 when not valid, instead of 0.
This guarantees that using them to index squares[] will throw an exception.