Eliminate duplicate positions in PossibleMates.
Sort the list of nonking pieces to force identical pieces
to be contiguous in the array. Then, eliminate duplicates
by forcing adjacent identical pieces to have strictly ordered
locations on the board.

For example, when given BB, the first bishop can be at c1
and the second bishop can be at d2, but not the other way around.