Figured out the most inclusive insufficient-material rule.
I updated Board.GetGameResult() with a simple rule that
terminates a game as an absolute draw as inclusively as possible.
PossibleMates confirms that any combination of more than one
of WN, WB, BN, BB on the board can produce a checkmate
through some legal series of moves.

Only when the sum of (WN, WB, BB, BN) <= 1 cannot possibly checkmate.