Tablebase: Refactored the planner logic as a class.
Now that I have the logic for planning the correct order
of generating endgame tables (to satisfy their dependencies
on each other), I moved that logic into a new class WorkPlanner.

The idea now is that I can pass in interfaces to do the actual
work of generating the tables, instead of just printing out the plan.