Starting to work on a function to calculate a table index from a chess board.
I will need to calculate an endgame table index from a chess board
in order to use the tables during a search.
This is my first attempt. The main problem is there are multiple
ways to calculate an index for the same position, due to nuances
of symmetry and interchangeable identical pieces.

So next I will have to go back and reconcile the indexes calculated
by the table generator and those calculated by the Board class.