Eliminated some redundant positions from the generator.
One tiny step closer to calculating the same indexes
in the table generator and in the Board class.
At the same time, I will be making the generator run faster
by computing fewer redundant positions.

When using 8-fold symmetry to restrict the White King to 10
distinct squares, there are redundancies with respect to the
placement of the Black King when the White King is on the
diagonal a1..d4. Then, flipping the board around the diagonal
a1..h8 changes the location of the Black King but not the
White King. The positions are equivalent because in both cases
we don't have any pawns on the board.

Before and after:
2020-12-23T00:51:49.37Z Found 4 checkmates. match=6364756, mismatch=167435931
2020-12-23T01:21:46.03Z Found 4 checkmates. match=6364756, mismatch=161005099