Refactored table index calculation to support flipping White/Black.
In the actual usage of the endgame tables, I will need to be able
to use the same table from either side's point of view.
This requires swapping Black/White pieces and flipping the offsets
so that White's side of the board becomes Black's.

Refactored the code so that the Position class contains most of the
logic for calculating the table index. The Board class just fills
in the Position class. This means it is easy to have the Board class
fill in the Position from either side's point of view.  Added
a reverseSides boolean parameter to Board.GetPosition(),
Board.GetEndgameConfig(), and Board.GetEndgameTableIndex().