Can evaluate two positions at once when there is White/Black material symmetry.
In material-symmetric configurations, when we establish the negamax score
for any position, we can immediately set the corresponding score for the
symmetric position. This should almost double the speed.

Even before this, there was a bug I just caught where we are counting
more positions than the table should be able to hold. I need to go back
and investigate that.

Renamed things to make the difference clearer between a configuration
and a configuration ID.

The function TableWorker.ConfigSymbol() was redundant.
Eliminate it and just use GetConfigId.ToString("D10") as needed.