Minor tweak to Board.ReverseSideOffset(int ofs).
If I ever use this function for a position that involves castling,
it will now work correctly. Before, it was flipping the board
around a horizontal axis. Now it rotates the board 180 degrees,
just like you were swapping sides with somebody. It should make
absolutely no difference for endgame generation, thanks to all
the symmetry rules.