Windows GUI: improve frame rate by limiting the rendering rectangle.
I was having trouble with the rendering keeping up with mouse
movement, and with it using a lot of CPU time also.

When animating dragging a piece, invalidate the smallest possible
rectangle to enclose its previous location and the current location.
When rendering, only draw the parts of the board that overlap
with the invalidated region.