Fixed #18 - Format an undefined score as "*.***".
When Thinker.Search() encounters a position in which there is
only one legal move, there is no point thinking about it.
It just immediately picks that move. It doesn't make sense
to assign any specific score to the move, so it just leaves
it as Score.Undefined = 0x80000000. Score.Format(), when given
this value, was returning a nonsensical string "#--523741824".
Now it returns "*.***" instead.

I thought about making it "?.???", but this is too similar
to the blunder notation "?" or "??"; I think asterisks are less confusing.