Starting to plan for generating endgame tables in parallel.
I'm working toward generating 5-piece endgame tables, which
will be more of a challenge. One thing that will help is to
generate the tables in parallel using multiple worker threads.

To do so requires computing the dependencies between the tables.
We need to know which configurations must be finished before we
start working on the next configuration in the work list.

Updated the "EndgameTableGen plan" command to compute the
dependencies between the generated configurations, as a
prelude to reworking as a parallel version.

I intend to create a new command "pgen" to do the parallel
computation, while keeping the single-threaded "gen" as a reference.