More efficient parallel endgame generator.
Instead of each worker thread picking a job and waiting
until it is ready to be worked, each thread now searches
through all jobs and picks the first one that is ready
to be worked.

Instead of busy-waiting with 1 second sleeps,
use a dedicated AutoResetEvent for the worker threads
to signal each other when another job has been completed.