Added DiskTable class for slower but less memory-intensive table sharing.
After a MemoryTable has been calculated, convert it to a DiskTable
that uses seek/read to load scores to assist calculating other tables.
The DiskTable can be "cloned" to allow multiple threads to independently
seek and read from the table.
This is part of my strategy of working toward 5-piece endgame tables.