Work in progress: overhauling the graph theory part of the algorithm to use less memory.
The endgame generator treats board positions as nodes in a graph,
and legal moves are edges between the nodes.
The old algorithm kept the nodes in memory, but that approach
will use too much memory to process 3 non-king pieces.

In addition to keeping the nodes in memory, the old algorithm
kept sweeping through all possible positions and finding what the
next positions are (in the same table or in prior tables).
Each time, it would add scores for the next layer n of mate-in-n-plies
positions.

I have what I believe will be a more efficient approach:
generate all the edges, but sort and index them by after-positions
(positions after making a move).

This commit is the first step. It writes binary files that contain
a list of edge records of the form

(
    before(config_id, white_to_move, table_index),
    after(config_id, white_to_move, table_index)
)

The goal is to sort these records by 'after', so that given
any after-position, we can find the set of all before-positions
in the current configuration leads to it via some legal move.