More endgame work in progress: radix sort edge files.
Implemented radix sort for the edge files by after_tindex.
Radix sort uses little memory and runs quickly.
The purpose is to pack all the edges with the same after_tindex
together in the file, so that given any after_tindex, we can
quickly find all the before_tindex that correspond to it.

I could use a binary search to find a given after_tindex,
but I think it will be simpler and faster to make an index
file on the very last iteration of the sort. Then I can seek
to the corresponding position in the index file, find the
offset of the first record with that after_tindex, and seek to it.

Then I can keep reading edges until I hit EOF or the after_tindex increases.