Creating indexes of the edge files.
To rapidly look up the edges for a given after_tindex,
the edge sorter now creates an index file.
Seeking to the location corresponding to the after_tindex
allows reading an integer telling where in the edge file
the first edge record is for that after_tindex.
The reader must then keep reading until the after_tindex
changes or hitting EOF. If the integer == -1, it means
there are no edges for that after_tindex (it is unreachable
or invalid).