Fixed #19 - NegaMax should treat known endgame positions as leaf nodes.
Thinker.NegaMax() was continuing to search deeper in quiescence search
when it found a position covered by endgame tables. It wasn't even
consulting the endgame tables in the full width search.

Now consult the endgame tables at all levels of the search
(full width and quiescence), and immediately return the score
for the position without generating legal moves or searching any deeper.