Work in progress: better endgame table unit tests.
Adding a more exhaustive endgame unit test, where I generate
all possible positions and compare full-width search with endgame tables,
but only when the full-width search is acceptably shallow compared
to the score returned by the endgame table.

Added class NullEvaluator, which is an evaluation function that always returns 0.
This way, the only time a nonzero score is returned by the search is when
it sees a forced win or a forced loss.
This new class will also be helpful for solving checkmate puzzles,
where material and position don't matter.

Thinker.Search() still immediately returns the move when there is only
one legal move for the given position. But now it checks to see if
it can assign an exact score to the move when the move it picks ends
the game, or has a known score in the endgame tables.
Otherwise, the returned move will still have Score.Undefined like
it did before.