Fixed bugs in calculating endgame table indexes due to multiple pieces of the same kind.
When there are N = 2, 3, ... pieces of the same kind,
there are N! different ways to place those pieces on the board
that create the same position. This leads to different table indexes
being calculated. To avoid this ambiguity, now I always keep
the list of pieces in ascending index order.

There was one tricky part having to do with figuring out which
symmetry transformation to pick. I have to examine diagonal heights
for identical pieces in the order of their post-transformed indexes,
even before applying the transform to the whole position.