More work on the endgame walker tests.
Check scores when the walker test finds
a terminal position: current player checkmated or definite draw.
The score must match.

Don't report an error when we can't find a child position in
the endgame database. This is normal, because some configurations
(for example, KB:kb) are excluded on purpose; checkmates are
possible but not forcible. Just assume such positions are a draw.

Updated reference sha256 hash for 0000100010.endgame, because
the backprop version passes a test that fails for the forward prop
version:

    Loaded 29 endgame tables from: /home/don/github/gearbox/tables/reference
    EndgameWalker[WB, WP] @ 11963049: best score 1099999939 (#31) [d7d8n] does not match parent score 0 (0.000) in 8/3P4/KBk5/8/8/8/8/8 w - - 0 1

I will need to go back and fix this in the forward prop.