Always recalculate ply horizon during each backprop ply level.
Just to be paranoid, it is so much simpler, and costs very
little extra time, to sweep through the whole bestScoreSoFar
table after each backprop ply, to guarantee that we have
the correct max_child_ply horizon value.  So I added a new
function FarthestPly() that does that.  This seems less
likely to have a bug in it than trying to maintain the max_child_ply.