Starting to work on endgame table data compression.
Work in progress. My first approach will be to use Huffman coding
of the scores, in blocks of 1K pairs of (White, Black).
There will be an offset table at the front for O(1) lookup
of each block, because the compressed blocks will be variable in size.
To access a score, the reader will seek to the correct slot in the
offset table, read the desired offset and the one after that.
Subtract to find total size of the compressed block.
Then seek to the block, read it, and decompress it.
Then access the desired score(s) inside the decompressed data.