Store block lengths instead of block offsets in compressed endgame tables.
This has two benefits:

1. The output is smaller because the block lengths use only 2 bytes
   per entry instead of 4.

2. It future-proofs us against any output files that may hold
   more than 4GB of data.

There is one downside: the decompressor will need to sweep through the
block lengths and calculate the positions of the blocks before any
random access of the blocks. It can do this one time and save the
block offset table in memory, so it seems like a small price to pay
for both of the benefits above.