Limit use of disk-based endgame tables to top level of the search tree.
The smaller endgame tables (1 or 2 nonking pieces) fit in memory,
and are therefore very efficient. However, the 3-nonking pieces
require disk I/O to consult. This change limits the use of these
slower tables to the very top level of the negamax search, to
prevent getting bogged down with disk I/O.